chore(wallet): Reusable Line Chart Controls
